@@ -3,7 +3,7 @@
     "type": "contao-bundle",
     "description": "Allows you to send your news as a newsletter using the Notification Center.",
     "keywords": ["contao", "news", "newsletter"],
-    "license": "LGPL-3.0+",
+    "license": "MIT",
     "authors": [
         {
             "name": "terminal42 gmbh",
@@ -26,15 +26,16 @@
         "forum": "https://community.contao.org"
     },
     "require": {
-        "php": "^7.1 || ^8.0",
-        "contao/core-bundle": "^4.9",
-        "contao/news-bundle": "^4.9",
-        "contao/newsletter-bundle": "^4.9",
-        "codefog/contao-haste": "^4.11.4 || ^5.0",
-        "terminal42/notification_center": "^1.3.5"
+        "php": "^8.1",
+        "contao/core-bundle": "^4.13 || ^5.0",
+        "contao/news-bundle": "^4.13 || ^5.0",
+        "contao/newsletter-bundle": "^4.13 || ^5.0",
+        "codefog/contao-haste": "^5.0",
+        "terminal42/notification_center": "^2.0"
     },
     "require-dev": {
-        "contao/manager-plugin": "^2.0"
+        "contao/manager-plugin": "^2.0",
+        "terminal42/contao-build-tools": "dev-main"
     },
     "conflict": {
         "contao/manager-plugin": "<2.0 || >=3.0"
@@ -46,8 +47,5 @@
     },
     "extra": {
         "contao-manager-plugin": "Terminal42\\NewsNewsletterBundle\\ContaoManager\\Plugin"
-    },
-    "scripts": {
-        "cs-fixer": "@php tools/ecs/vendor/bin/ecs check src/ --config tools/ecs/config.php --fix --ansi"
     }
 }
