@@ -1,7 +1,3 @@
 <?php
 
-/**
- * Notification types
- */
-$GLOBALS['TL_LANG']['tl_nc_notification']['type']['news_newsletter']         = 'News Newsletter';
-$GLOBALS['TL_LANG']['tl_nc_notification']['type']['news_newsletter_default'] = array('Standard', 'Dieser Benachrichtigungstyp kann für Newsletter bei Veröffentlichung eines News-Beitrags verwendet werden.');
+$GLOBALS['TL_LANG']['tl_nc_notification']['type']['news_newsletter'] = ['News Newsletter', 'Dieser Benachrichtigungstyp kann für Newsletter bei Veröffentlichung eines News-Beitrags verwendet werden.'];
