@@ -1,18 +1,9 @@
 <?php
 
-/**
- * Fields
- */
-$GLOBALS['TL_LANG']['tl_news']['newsletter'] = array('Newsletter wurde verschickt');
+$GLOBALS['TL_LANG']['tl_news']['newsletter'] = ['Newsletter wurde verschickt'];
 
-/**
- * Buttons
- */
-$GLOBALS['TL_LANG']['tl_news']['sendNewsletter']        = array('Newsletter senden', 'Newsletter für News ID %s senden');
+$GLOBALS['TL_LANG']['tl_news']['sendNewsletter'] = ['Newsletter senden', 'Newsletter für News ID %s senden'];
 $GLOBALS['TL_LANG']['tl_news']['sendNewsletterConfirm'] = 'Möchten Sie den Newsletter wirklich an %s Empfänger senden?';
 
-/**
- * Messages
- */
 $GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_confirm'] = 'Newsletter wurde erfolgreich verschickt!';
-$GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_error']   = 'Newsletter konnte nicht versendet werden.';
+$GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_error'] = 'Newsletter konnte nicht versendet werden.';
