@@ -1,12 +1,6 @@
 <?php
 
-/**
- * Fields
- */
-$GLOBALS['TL_LANG']['tl_news_archive']['newsletter']         = array('Newsletter aktivieren', 'Newsletter für Neuigkeiten aktivieren.');
-$GLOBALS['TL_LANG']['tl_news_archive']['newsletter_channel'] = array('Newsletter-Verteiler', 'Bitte wählen Sie hier den Verteiler.');
+$GLOBALS['TL_LANG']['tl_news_archive']['newsletter'] = ['Newsletter aktivieren', 'Newsletter für Neuigkeiten aktivieren.'];
+$GLOBALS['TL_LANG']['tl_news_archive']['newsletter_channel'] = ['Newsletter-Verteiler', 'Bitte wählen Sie hier den Verteiler.'];
 
-/**
- * Legends
- */
 $GLOBALS['TL_LANG']['tl_news_archive']['newsletter_legend'] = 'Newsletter-Einstellungen';
