@@ -1,7 +1,3 @@
 <?php
 
-/**
- * Notification types
- */
-$GLOBALS['TL_LANG']['tl_nc_notification']['type']['news_newsletter']         = 'News newsletter';
-$GLOBALS['TL_LANG']['tl_nc_notification']['type']['news_newsletter_default'] = array('Default', 'This notification type can be sent when the news article is published.');
+$GLOBALS['TL_LANG']['tl_nc_notification']['type']['news_newsletter'] = ['News newsletter', 'This notification type can be sent when the news article is published.'];
