@@ -1,18 +1,9 @@
 <?php
 
-/**
- * Fields
- */
-$GLOBALS['TL_LANG']['tl_news']['newsletter'] = array('Newsletter sent');
+$GLOBALS['TL_LANG']['tl_news']['newsletter'] = ['Newsletter sent'];
 
-/**
- * Buttons
- */
-$GLOBALS['TL_LANG']['tl_news']['sendNewsletter']        = array('Send newsletter', 'Send the newsletter for news ID %s');
+$GLOBALS['TL_LANG']['tl_news']['sendNewsletter'] = ['Send newsletter', 'Send the newsletter for news ID %s'];
 $GLOBALS['TL_LANG']['tl_news']['sendNewsletterConfirm'] = 'Do you really want to send the newsletter to %s recipients?';
 
-/**
- * Messages
- */
 $GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_confirm'] = 'Newsletter sent successfully!';
-$GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_error']   = 'Could not send the newsletter.';
+$GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_error'] = 'Could not send the newsletter.';
