@@ -1,12 +1,6 @@
 <?php
 
-/**
- * Fields
- */
-$GLOBALS['TL_LANG']['tl_news_archive']['newsletter']         = array('Enable newsletter', 'Enable the newsletter to send news articles.');
-$GLOBALS['TL_LANG']['tl_news_archive']['newsletter_channel'] = array('Newsletter channel', 'Please choose the newsletter channel.');
+$GLOBALS['TL_LANG']['tl_news_archive']['newsletter'] = ['Enable newsletter', 'Enable the newsletter to send news articles.'];
+$GLOBALS['TL_LANG']['tl_news_archive']['newsletter_channel'] = ['Newsletter channel', 'Please choose the newsletter channel.'];
 
-/**
- * Legends
- */
 $GLOBALS['TL_LANG']['tl_news_archive']['newsletter_legend'] = 'Newsletter settings';
