@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Terminal42\NewsNewsletterBundle\ContaoManager;
 
+use Codefog\HasteBundle\CodefogHasteBundle;
 use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
@@ -11,6 +12,7 @@ use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 use Contao\NewsBundle\ContaoNewsBundle;
 use Contao\NewsletterBundle\ContaoNewsletterBundle;
 use Terminal42\NewsNewsletterBundle\Terminal42NewsNewsletterBundle;
+use Terminal42\NotificationCenterBundle\Terminal42NotificationCenterBundle;
 
 class Plugin implements BundlePluginInterface
 {
@@ -22,8 +24,8 @@ class Plugin implements BundlePluginInterface
                     ContaoCoreBundle::class,
                     ContaoNewsBundle::class,
                     ContaoNewsletterBundle::class,
-                    'haste',
-                    'notification_center',
+                    CodefogHasteBundle::class,
+                    Terminal42NotificationCenterBundle::class,
                 ])
                 ->setReplace(['news_newsletter']),
         ];
