@@ -8,8 +8,8 @@ use Codefog\HasteBundle\Formatter;
 use Contao\Backend;
 use Contao\ContentModel;
 use Contao\Controller;
+use Contao\CoreBundle\DependencyInjection\Attribute\AsCallback;
 use Contao\CoreBundle\Exception\RedirectResponseException;
-use Contao\CoreBundle\ServiceAnnotation\Callback;
 use Contao\Image;
 use Contao\Input;
 use Contao\Message;
@@ -19,14 +19,24 @@ use Contao\NewsModel;
 use Contao\PageModel;
 use Contao\StringUtil;
 use Contao\System;
-use Haste\Util\Format;
-use NotificationCenter\Model\Notification;
+use Terminal42\NotificationCenterBundle\Exception\Parcel\CouldNotCreateParcelException;
+use Terminal42\NotificationCenterBundle\NotificationCenter;
 
-/**
- * @Callback(table="tl_news", target="list.operations.newsletter.button")
- */
+#[AsCallback(table: 'tl_news', target: 'list.operations.newsletter.button')]
 class SendNewsletterListener
 {
+    public function __construct(
+        private readonly NotificationCenter $notificationCenter,
+        private readonly Formatter $formatter,
+    ) {
+    }
+
+    /**
+     * @param string|null $href
+     * @param string|null $label
+     * @param string|null $title
+     * @param string|null $icon
+     */
     public function __invoke(array $row, $href, $label, $title, $icon): string
     {
         $archive = NewsArchiveModel::findByPk($row['pid']);
@@ -37,7 +47,7 @@ class SendNewsletterListener
 
         // Toggle the record
         if (Input::get('newsletter')) {
-            if ($this->sendNewsMessage(Input::get('newsletter'))) {
+            if ($this->sendNewsMessage((int) Input::get('newsletter'))) {
                 Message::addConfirmation($GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_confirm']);
             } else {
                 Message::addError($GLOBALS['TL_LANG']['tl_news']['message_news_newsletter_error']);
@@ -48,7 +58,7 @@ class SendNewsletterListener
 
         // Return just an image if newsletter was sent
         if ($row['newsletter']) {
-            return Image::getHtml(str_replace('.png', '_.png', $icon), $label);
+            return Image::getHtml(str_replace('.png', '_.png', (string) $icon), $label);
         }
 
         // Add the confirmation popup
@@ -58,78 +68,66 @@ class SendNewsletterListener
         return '<a href="'.Backend::addToUrl($href.'&newsletter='.$row['id']).'" title="'.StringUtil::specialchars($title).'"'.$attributes.'>'.Image::getHtml($icon, $label).'</a> ';
     }
 
-    private function sendNewsMessage($intId)
+    private function sendNewsMessage(int $newsletterId): bool
     {
-        $objNews = NewsModel::findByPk($intId);
+        $news = NewsModel::findByPk($newsletterId);
 
-        if (null === $objNews) {
+        if (null === $news) {
             return false;
         }
 
-        $objArchive = $objNews->getRelated('pid');
+        $newsArchive = $news->getRelated('pid');
 
-        if (null === $objArchive || !$objArchive->newsletter || !$objArchive->newsletter_channel || !$objArchive->nc_notification) {
+        if (null === $newsArchive || !$newsArchive->newsletter || !$newsArchive->newsletter_channel || !$newsArchive->nc_notification) {
             return false;
         }
 
-        $objNotification = Notification::findByPk($objArchive->nc_notification);
+        $recipients = NewsletterRecipientsModel::findBy(['pid=? AND active=1'], $newsArchive->newsletter_channel);
 
-        if (null === $objNotification) {
+        if (null === $recipients) {
             return false;
         }
 
-        $objRecipients = NewsletterRecipientsModel::findBy(['pid=? AND active=1'], $objArchive->newsletter_channel);
-
-        if (null === $objRecipients) {
-            return false;
-        }
-
-        $arrTokens = [];
+        $tokens = [];
 
         // Generate news archive tokens
-        foreach ($objArchive->row() as $k => $v) {
-            if (class_exists(Format::class)) {
-                $arrTokens['news_archive_'.$k] = Format::dcaValue('tl_news_archive', $k, $v);
-            } else {
-                $arrTokens['news_archive_'.$k] = System::getContainer()->get(Formatter::class)->dcaValue('tl_news_archive', $k, $v);
-            }
+        foreach ($newsArchive->row() as $k => $v) {
+            $tokens['news_archive_'.$k] = $this->formatter->dcaValue('tl_news_archive', $k, $v);
         }
 
         // Generate news tokens
-        foreach ($objNews->row() as $k => $v) {
-            if (class_exists(Format::class)) {
-                $arrTokens['news_'.$k] = Format::dcaValue('tl_news', $k, $v);
-            } else {
-                $arrTokens['news_'.$k] = System::getContainer()->get(Formatter::class)->dcaValue('tl_news', $k, $v);
-            }
+        foreach ($news->row() as $k => $v) {
+            $tokens['news_'.$k] = $this->formatter->dcaValue('tl_news', $k, $v);
         }
 
-        $arrTokens['news_text'] = '';
-        $objElement = ContentModel::findPublishedByPidAndTable($objNews->id, 'tl_news');
+        $tokens['news_text'] = '';
+        $elements = ContentModel::findPublishedByPidAndTable($news->id, 'tl_news');
 
         // Generate news text
-        if (null !== $objElement) {
-            while ($objElement->next()) {
-                $arrTokens['news_text'] .= Controller::getContentElement($objElement->id);
+        if (null !== $elements) {
+            foreach ($elements as $element) {
+                $tokens['news_text'] .= Controller::getContentElement($element->id);
             }
         }
 
         // Generate news URL
-        if (null !== ($objPage = PageModel::findWithDetails($objNews->getRelated('pid')->jumpTo))) {
-            $arrTokens['news_url'] = $objPage->getAbsoluteUrl(($GLOBALS['TL_CONFIG']['useAutoItem'] ? '/' : '/items/').($objNews->alias ?: $objNews->id));
+        if (null !== ($objPage = PageModel::findWithDetails($news->getRelated('pid')->jumpTo))) {
+            $tokens['news_url'] = $objPage->getAbsoluteUrl('/'.($news->alias ?: $news->id));
         }
 
-        // Administrator e-mail
-        $arrTokens['admin_email'] = $GLOBALS['TL_ADMIN_EMAIL'];
+        foreach ($recipients as $recipient) {
+            $tokens['recipient_email'] = $recipient->email;
 
-        while ($objRecipients->next()) {
-            $arrTokens['recipient_email'] = $objRecipients->email;
-            $objNotification->send($arrTokens);
+            try {
+                $this->notificationCenter->sendNotification($newsArchive->nc_notification, $tokens);
+            } catch (CouldNotCreateParcelException) {
+                return false;
+            }
         }
 
         // Set the newsletter flag
-        $objNews->newsletter = 1;
-        $objNews->save();
+        $news->newsletter = 1;
+        $news->save();
 
         return true;
     }
